api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
groq==0.4.1
transformers==4.36.2
torch==2.1.2
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0